% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_scientific}
\alias{format_scientific}
\title{Format Scientific Notation}
\usage{
format_scientific(x, digits = 3)
}
\arguments{
\item{x}{Numeric vector to format.}

\item{digits}{Number of significant digits to display.}
}
\value{
Character vector of formatted numbers.
}
\description{
Format numbers in scientific notation when the absolute value is below 0.001,
otherwise in fixed notation.
}
