% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_config}
\alias{get_config}
\title{Get Configuration Value}
\usage{
get_config(key, default = NULL)
}
\arguments{
\item{key}{Configuration key.}

\item{default}{Default value if the key is not found.}
}
\value{
Configuration value or default.
}
\description{
Retrieve a configuration value from the global \code{iQualityR.config} option,
with a fallback default.
}
