% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i18n.R
\name{iqr_locale}
\alias{iqr_locale}
\title{Get the active iQualityR locale}
\usage{
iqr_locale(locale = NULL)
}
\arguments{
\item{locale}{Optional locale code. Supported values are \code{"en-US"} and \code{"zh-CN"}.}
}
\value{
Locale code.
}
\description{
Get the active iQualityR locale
}
