% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i18n.R
\name{iqr_set_locale}
\alias{iqr_set_locale}
\title{Set the active iQualityR locale}
\usage{
iqr_set_locale(locale = c("en-US", "zh-CN"))
}
\arguments{
\item{locale}{Locale code.}
}
\value{
The selected locale, invisibly.
}
\description{
Set the active iQualityR locale
}
