% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IqrAnalyzerBase.R
\name{IqrAnalyzerBase}
\alias{IqrAnalyzerBase}
\title{iQualityR Analyzer Base Class}
\description{
Abstract executor base class. Responsible for defining result storage protocol and handling parameter normalization.
Follows v2.0 framework specification: supports Task-Executor matrix architecture.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{results}}{Standardized result container (list with statistics, diagnostics, data_tables, raw_output).}

\item{\code{params}}{Named list of parameters (after setup).}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-IqrAnalyzerBase-new}{\code{IqrAnalyzerBase$new()}}
\item \href{#method-IqrAnalyzerBase-reset}{\code{IqrAnalyzerBase$reset()}}
\item \href{#method-IqrAnalyzerBase-set_statistic}{\code{IqrAnalyzerBase$set_statistic()}}
\item \href{#method-IqrAnalyzerBase-set_diagnostic}{\code{IqrAnalyzerBase$set_diagnostic()}}
\item \href{#method-IqrAnalyzerBase-set_datatable}{\code{IqrAnalyzerBase$set_datatable()}}
\item \href{#method-IqrAnalyzerBase-set_raw_output}{\code{IqrAnalyzerBase$set_raw_output()}}
\item \href{#method-IqrAnalyzerBase-setup}{\code{IqrAnalyzerBase$setup()}}
\item \href{#method-IqrAnalyzerBase-run}{\code{IqrAnalyzerBase$run()}}
\item \href{#method-IqrAnalyzerBase-get_results}{\code{IqrAnalyzerBase$get_results()}}
\item \href{#method-IqrAnalyzerBase-clone}{\code{IqrAnalyzerBase$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IqrAnalyzerBase-new"></a>}}
\if{latex}{\out{\hypertarget{method-IqrAnalyzerBase-new}{}}}
\subsection{Method \code{new()}}{
Initialize standardized result container.
Container structure is aligned with the needs of Reporter and Plotter.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IqrAnalyzerBase$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IqrAnalyzerBase-reset"></a>}}
\if{latex}{\out{\hypertarget{method-IqrAnalyzerBase-reset}{}}}
\subsection{Method \code{reset()}}{
Reset result container to empty state.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IqrAnalyzerBase$reset()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IqrAnalyzerBase-set_statistic"></a>}}
\if{latex}{\out{\hypertarget{method-IqrAnalyzerBase-set_statistic}{}}}
\subsection{Method \code{set_statistic()}}{
Store a single statistic in the results container.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IqrAnalyzerBase$set_statistic(key, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{Character. Name of the statistic.}

\item{\code{value}}{Value of the statistic.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IqrAnalyzerBase-set_diagnostic"></a>}}
\if{latex}{\out{\hypertarget{method-IqrAnalyzerBase-set_diagnostic}{}}}
\subsection{Method \code{set_diagnostic()}}{
Store a single diagnostic entry in the results container.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IqrAnalyzerBase$set_diagnostic(key, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{Character. Name of the diagnostic.}

\item{\code{value}}{Value of the diagnostic.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IqrAnalyzerBase-set_datatable"></a>}}
\if{latex}{\out{\hypertarget{method-IqrAnalyzerBase-set_datatable}{}}}
\subsection{Method \code{set_datatable()}}{
Store a single data table in the results container.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IqrAnalyzerBase$set_datatable(key, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{Character. Name of the data table.}

\item{\code{value}}{Data frame to store.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IqrAnalyzerBase-set_raw_output"></a>}}
\if{latex}{\out{\hypertarget{method-IqrAnalyzerBase-set_raw_output}{}}}
\subsection{Method \code{set_raw_output()}}{
Store raw output from an analysis.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IqrAnalyzerBase$set_raw_output(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{Arbitrary R object returned by the underlying analysis.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IqrAnalyzerBase-setup"></a>}}
\if{latex}{\out{\hypertarget{method-IqrAnalyzerBase-setup}{}}}
\subsection{Method \code{setup()}}{
Set analyzer parameters from a plan object or a list.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IqrAnalyzerBase$setup(input = list())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{An \code{IqrPlanBase} object or a named list.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IqrAnalyzerBase-run"></a>}}
\if{latex}{\out{\hypertarget{method-IqrAnalyzerBase-run}{}}}
\subsection{Method \code{run()}}{
Core entry point for executing analysis.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IqrAnalyzerBase$run(data, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Observation data. Accepts data.frame, data.table, tibble, or matrix.
data.frame-like objects are passed through unchanged; matrices are coerced to data.frame.}

\item{\code{...}}{Additional named parameters merged into \code{params}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IqrAnalyzerBase-get_results"></a>}}
\if{latex}{\out{\hypertarget{method-IqrAnalyzerBase-get_results}{}}}
\subsection{Method \code{get_results()}}{
Retrieve standardized results.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IqrAnalyzerBase$get_results()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IqrAnalyzerBase-clone"></a>}}
\if{latex}{\out{\hypertarget{method-IqrAnalyzerBase-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IqrAnalyzerBase$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
