% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IqrTheme.R
\name{IqrTheme}
\alias{IqrTheme}
\title{iQualityR Theme Class}
\description{
Theme management class for iQualityR, providing unified access to theme configuration,
plot styling, and export functionality.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{config}}{ThemeConfig object.}

\item{\code{plot}}{PlotTheme object.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-IqrTheme-new}{\code{IqrTheme$new()}}
\item \href{#method-IqrTheme-set_theme}{\code{IqrTheme$set_theme()}}
\item \href{#method-IqrTheme-get_pal}{\code{IqrTheme$get_pal()}}
\item \href{#method-IqrTheme-get_ui_colors}{\code{IqrTheme$get_ui_colors()}}
\item \href{#method-IqrTheme-get_data_colors}{\code{IqrTheme$get_data_colors()}}
\item \href{#method-IqrTheme-scale_fill_iqr}{\code{IqrTheme$scale_fill_iqr()}}
\item \href{#method-IqrTheme-scale_color_iqr}{\code{IqrTheme$scale_color_iqr()}}
\item \href{#method-IqrTheme-theme_iqr}{\code{IqrTheme$theme_iqr()}}
\item \href{#method-IqrTheme-clone}{\code{IqrTheme$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IqrTheme-new"></a>}}
\if{latex}{\out{\hypertarget{method-IqrTheme-new}{}}}
\subsection{Method \code{new()}}{
Create a new IqrTheme instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IqrTheme$new(theme_style = "academic", ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{theme_style}}{Character. Theme style (built-in "academic", "tech", or external name/function).}

\item{\code{...}}{Additional arguments passed to ThemeConfig$new().}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IqrTheme-set_theme"></a>}}
\if{latex}{\out{\hypertarget{method-IqrTheme-set_theme}{}}}
\subsection{Method \code{set_theme()}}{
Set the active theme (delegated to ThemeConfig).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IqrTheme$set_theme(style, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{style}}{Character or function.}

\item{\code{...}}{Additional arguments passed to ThemeConfig$set_theme().}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IqrTheme-get_pal"></a>}}
\if{latex}{\out{\hypertarget{method-IqrTheme-get_pal}{}}}
\subsection{Method \code{get_pal()}}{
Get color palette (delegated to ThemeConfig).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IqrTheme$get_pal(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Arguments passed to ThemeConfig$get_pal().}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IqrTheme-get_ui_colors"></a>}}
\if{latex}{\out{\hypertarget{method-IqrTheme-get_ui_colors}{}}}
\subsection{Method \code{get_ui_colors()}}{
Get UI theme colors.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IqrTheme$get_ui_colors()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IqrTheme-get_data_colors"></a>}}
\if{latex}{\out{\hypertarget{method-IqrTheme-get_data_colors}{}}}
\subsection{Method \code{get_data_colors()}}{
Get data colors.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IqrTheme$get_data_colors(type = c("discrete", "continuous"), custom = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{Character. Either "discrete" or "continuous".}

\item{\code{custom}}{Optional custom color vector to use instead.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IqrTheme-scale_fill_iqr"></a>}}
\if{latex}{\out{\hypertarget{method-IqrTheme-scale_fill_iqr}{}}}
\subsection{Method \code{scale_fill_iqr()}}{
Create ggplot2 fill scale (delegated to PlotTheme).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IqrTheme$scale_fill_iqr(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Arguments passed to PlotTheme$scale_fill_iqr().}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IqrTheme-scale_color_iqr"></a>}}
\if{latex}{\out{\hypertarget{method-IqrTheme-scale_color_iqr}{}}}
\subsection{Method \code{scale_color_iqr()}}{
Create ggplot2 color scale (delegated to PlotTheme).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IqrTheme$scale_color_iqr(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Arguments passed to PlotTheme$scale_color_iqr().}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IqrTheme-theme_iqr"></a>}}
\if{latex}{\out{\hypertarget{method-IqrTheme-theme_iqr}{}}}
\subsection{Method \code{theme_iqr()}}{
Create ggplot2 theme (delegated to PlotTheme).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IqrTheme$theme_iqr(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Arguments passed to PlotTheme$theme_iqr().}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IqrTheme-clone"></a>}}
\if{latex}{\out{\hypertarget{method-IqrTheme-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IqrTheme$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
