% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IqrTheme.R
\name{mix}
\alias{mix}
\title{Mix two colors}
\usage{
mix(color1, color2, amount = 0.5)
}
\arguments{
\item{color1}{Character. A hex color string.}

\item{color2}{Character. A hex color string.}

\item{amount}{Numeric between 0 and 1. Proportion of \code{color2} in the result.}
}
\value{
Character. A hex color string.
}
\description{
Linearly interpolate between two colors.
}
\examples{
mix("#FF0000", "#0000FF", 0.5)
}
