% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{null_to_na}
\alias{null_to_na}
\title{Convert NULL to NA}
\usage{
null_to_na(x)
}
\arguments{
\item{x}{Value to convert.}
}
\value{
NA if \code{x} is NULL, otherwise \code{x} unchanged.
}
\description{
Simple helper to convert NULL values to NA.
}
