% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_metadata}
\alias{validate_metadata}
\title{Validate Metadata}
\usage{
validate_metadata(metadata)
}
\arguments{
\item{metadata}{List containing metadata.}
}
\value{
The validated metadata list with all required category slots initialized.
}
\description{
Validate and ensure the presence of required category slots in a metadata list
(man, machine, material, method, environment, project).
}
