% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IqrPlotterBase.R
\name{IqrPlotterBase}
\alias{IqrPlotterBase}
\title{iQualityR Plotter Base Class}
\description{
Abstract plotter base class. Defines the interface for rendering plots.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-IqrPlotterBase-render}{\code{IqrPlotterBase$render()}}
\item \href{#method-IqrPlotterBase-clone}{\code{IqrPlotterBase$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IqrPlotterBase-render"></a>}}
\if{latex}{\out{\hypertarget{method-IqrPlotterBase-render}{}}}
\subsection{Method \code{render()}}{
Render plot (virtual method, must be implemented by subclass).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IqrPlotterBase$render(results, theme_obj, type = "full", ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{results}}{Analysis results (list from Analyzer).}

\item{\code{theme_obj}}{IqrTheme object.}

\item{\code{type}}{Character. Plot type selector, default "full".}

\item{\code{...}}{Additional parameters passed to subclass implementations.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IqrPlotterBase-clone"></a>}}
\if{latex}{\out{\hypertarget{method-IqrPlotterBase-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IqrPlotterBase$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
